Cursor and Physics Minigame
This is a commit of my cursor that changes colored sprites when clicked and released. Using the cursor you can also drag an item that will affect the physics of the other objects when bumped into them.

Prefabs made: Physics Item and Cursor.

The camera needs to be orthographic for the minigame to work!

Script files made in this commit: PhysicsObject and ChangeCursorOnClick.